add LessonStatus

+++ src/Command/InitDatabaseCommand.php
@@ -3,6 +3,7 @@
 namespace App\Command;
 
 use App\DataFixtures\Factory\UserFactory;
+use App\Entity\Lesson\LessonStatus;
 use App\Entity\SchoolClass\SchoolClassStatus;
 use App\Enum\UserType;
 use Doctrine\ORM\EntityManagerInterface;
@@ -33,6 +34,7 @@ class InitDatabaseCommand extends Command
 
         $this->createOwner();
         $this->initSchoolClassStatuses();
+        $this->initLessonStatuses();
 
         $io->success('OK');
         return Command::SUCCESS;
@@ -52,11 +54,49 @@ class InitDatabaseCommand extends Command
 
     private function initSchoolClassStatuses()
     {
-        $this->entityManager->createQuery('DELETE FROM App\Entity\SchoolClass\SchoolClassStatus')->execute();
-        $this->entityManager->getConnection()->exec("ALTER TABLE school_class_status AUTO_INCREMENT = 1");
-
         $repository = $this->entityManager->getRepository(SchoolClassStatus::class);
-        $repository->add((new SchoolClassStatus())->setName(SchoolClassStatus::ACTIVE));
-        $repository->add((new SchoolClassStatus())->setName(SchoolClassStatus::ARCHIVED), true);
+
+        if (empty($repository->findAll())) {
+            $this->entityManager->createQuery('DELETE FROM App\Entity\SchoolClass\SchoolClassStatus')->execute();
+            $this->entityManager->getConnection()->exec("ALTER TABLE school_class_status AUTO_INCREMENT = 1");
+
+
+            $repository->add((new SchoolClassStatus())->setName(SchoolClassStatus::ACTIVE));
+            $repository->add((new SchoolClassStatus())->setName(SchoolClassStatus::ARCHIVED), true);
+        }
+    }
+
+    private function initLessonStatuses()
+    {
+        $repository = $this->entityManager->getRepository(LessonStatus::class);
+
+        if (empty($repository->findAll())) {
+
+            $this->entityManager->createQuery('DELETE FROM App\Entity\Lesson\LessonStatus')->execute();
+            $this->entityManager->getConnection()->exec("ALTER TABLE lesson_status AUTO_INCREMENT = 1");
+
+            $data = [
+                [
+                    "id" => 1,
+                    "description" => "zgodnie z planem",
+                    "background" => "bg-success"
+                ],
+                [
+                    "id" => 2,
+                    "description" => "odwołane",
+                    "background" => "bg-danger"
+                ],
+                [
+                    "id" => 3,
+                    "description" => "zmodyfikowane",
+                    "background" => "bg-info"
+                ]
+            ];
+
+            foreach ($data as $index => $item) {
+                $lessonStatus = new LessonStatus($item['id'], $item['description'], $item['background']);
+                $repository->add($lessonStatus, $index + 1 == count($data));
+            }
+        }
     }
 }
